@@ -77,8 +77,8 @@ namespace Apache.Arrow.Adbc.Drivers.Apache
         public const string ForeignTableName = "adbc.get_metadata.foreign_target_table";
 
         /// <summary>
-        /// Whether to escape underscore to treat as literal rather than wildcard. Default to false.
+        /// Whether to escape pattern wildcard characters (_ and %) to treat as literal rather than wildcard. Default to false.
         /// </summary>
-        public const string EscapeUnderscore = "adbc.get_metadata.escape_underscore";
+        public const string EscapePatternWildcards = "adbc.get_metadata.escape_pattern_wildcards";
     }
 }
