@@ -159,12 +159,12 @@ namespace Apache.Arrow.Adbc.Tests.Drivers.Interop.FlightSql
         {
             foreach (FlightSqlTestEnvironment environment in _environments)
             {
-                environment.Query = "SELECT * WHERE 0=1";
+                environment.Query = "SELECT 1 WHERE 0=1"; // The query simulates no results
                 environment.ExpectedResultsCount = 0;
 
                 using (Adbc.Client.AdbcConnection adbcConnection = GetFlightSqlAdbcConnectionUsingConnectionString(environment, _testConfiguration))
                 {
-                    Tests.ClientTests.CanClientExecuteQuery(adbcConnection, environment, additionalCommandOptionsSetter: null, environment.Name);
+                    Tests.ClientTests.CanClientExecuteQuery(adbcConnection, environment, additionalCommandOptionsSetter: null, environmentName: environment.Name);
                 }
             }
         }
