@@ -36,12 +36,19 @@ var (
 
 func init() {
 	if info, ok := debug.ReadBuildInfo(); ok {
+		for _, s := range info.Settings {
+			switch s.Key {
+			case "vcs.modified":
+				if s.Value == "true" {
+					infoDriverVersion += "-dev"
+				}
+			}
+		}
 		for _, dep := range info.Deps {
 			switch {
-			case dep.Path == "github.com/apache/arrow-adbc/go/adbc":
-				infoDriverVersion = dep.Version
-			case strings.HasPrefix(dep.Path, "github.com/apache/arrow/go/"):
+			case strings.HasPrefix(dep.Path, "github.com/apache/arrow-go/"):
 				infoDriverArrowVersion = dep.Version
+				return
 			}
 		}
 	}
