@@ -20,11 +20,13 @@ package drivermgr
 // #cgo !windows LDFLAGS: -ldl
 // #cgo windows CFLAGS: -DADBC_EXPORTING
 // #cgo windows CPPFLAGS: -DADBC_EXPORTING
-// #cgo CXXFLAGS: -std=c++11
+// #cgo windows LDFLAGS: -lshell32 -ladvapi32 -luuid
+// #cgo CXXFLAGS: -std=c++17 -I${SRCDIR}/../../../c/vendor/
 // #if !defined(ADBC_EXPORTING)
 // #define ADBC_EXPORTING
 // #endif
 // #include "arrow-adbc/adbc.h"
+// #include "arrow-adbc/adbc_driver_manager.h"
 // #include <stdlib.h>
 // #include <string.h>
 //
@@ -49,6 +51,7 @@ package drivermgr
 import "C"
 import (
 	"context"
+	"strconv"
 	"sync"
 	"unsafe"
 
@@ -56,6 +59,18 @@ import (
 	"github.com/apache/arrow-go/v18/arrow"
 	"github.com/apache/arrow-go/v18/arrow/array"
 	"github.com/apache/arrow-go/v18/arrow/cdata"
+)
+
+const (
+	LoadFlagsSearchEnv = 1 << iota
+	LoadFlagsSearchPath
+	LoadFlagsSearchSystem
+	LoadFlagsAllowRelativePaths
+
+	LoadFlagsDefault = LoadFlagsSearchEnv | LoadFlagsSearchPath | LoadFlagsSearchSystem | LoadFlagsAllowRelativePaths
+	// LoadFlagsOptionKey is the key to use for an option to set specific
+	// load flags for the database to decide where to look for driver manifests.
+	LoadFlagsOptionKey = "load_flags"
 )
 
 type option struct {
@@ -92,18 +107,54 @@ func (d Driver) NewDatabaseWithContext(_ context.Context, opts map[string]string
 		options: make(map[string]option),
 	}
 
+	defer func() {
+		if db.db == nil { // cleanup options if we failed to create the database
+			for _, o := range dbOptions {
+				C.free(unsafe.Pointer(o.key))
+				C.free(unsafe.Pointer(o.val))
+			}
+		}
+	}()
+
 	var err C.struct_AdbcError
 	db.db = (*C.struct_AdbcDatabase)(unsafe.Pointer(C.calloc(C.sizeof_struct_AdbcDatabase, C.size_t(1))))
 	if code := adbc.Status(C.AdbcDatabaseNew(db.db, &err)); code != adbc.StatusOK {
 		return nil, toAdbcError(code, &err)
 	}
 
-	for _, o := range dbOptions {
-		if code := adbc.Status(C.AdbcDatabaseSetOption(db.db, o.key, o.val, &err)); code != adbc.StatusOK {
-			errOut := toAdbcError(code, &err)
-			C.AdbcDatabaseRelease(db.db, &err)
-			db.db = nil
-			return nil, errOut
+	if code := adbc.Status(C.AdbcDriverManagerDatabaseSetLoadFlags(db.db, C.AdbcLoadFlags(LoadFlagsDefault), &err)); code != adbc.StatusOK {
+		errOut := toAdbcError(code, &err)
+		C.AdbcDatabaseRelease(db.db, &err)
+		db.db = nil
+		return nil, errOut
+	}
+
+	for k, o := range dbOptions {
+		switch k {
+		case LoadFlagsOptionKey:
+			f, errOut := strconv.Atoi(C.GoString(o.val))
+			if errOut != nil {
+				C.AdbcDatabaseRelease(db.db, &err)
+				db.db = nil
+				return nil, adbc.Error{
+					Code: adbc.StatusInvalidArgument,
+					Msg:  "invalid load flags value: " + C.GoString(o.val),
+				}
+			}
+
+			if code := adbc.Status(C.AdbcDriverManagerDatabaseSetLoadFlags(db.db, C.AdbcLoadFlags(f), &err)); code != adbc.StatusOK {
+				errOut := toAdbcError(code, &err)
+				C.AdbcDatabaseRelease(db.db, &err)
+				db.db = nil
+				return nil, errOut
+			}
+		default:
+			if code := adbc.Status(C.AdbcDatabaseSetOption(db.db, o.key, o.val, &err)); code != adbc.StatusOK {
+				errOut := toAdbcError(code, &err)
+				C.AdbcDatabaseRelease(db.db, &err)
+				db.db = nil
+				return nil, errOut
+			}
 		}
 	}
 
